fix: the improper UI regarding the members amount is properly handled
In this commit you can see the partially fixed first part of the bug. Now, It properly resolves the API about the users amount and shows the right amount of members with the properly handled logic about how many of them should be displayed and how. But why partially? Because by this API there are some datas which should be added in the future, one of them is the icons url, but now I already added the future proof solution for that. I just need to have some additional communication with our BE team (Serhat specifically) regarding to add the same icon property with the proper naming convention which matches the front-end one in the types file.