#Ng-379 (#447)
* fix: the improper UI regarding the members amount is properly handled

In this commit you can see the partially fixed first part of the bug. Now, It properly resolves the API about the users amount and shows the right amount of members with the properly handled logic about how many of them should be displayed and how. But why partially? Because by this API there are some datas which should be added in the future, one of them is the icons url, but now I already added the future proof solution for that. I just need to have some additional communication with our BE team (Serhat specifically) regarding to add the same icon property with the proper naming convention which matches the front-end one in the types file.

* fix: experiment amount displays properly

After the communication with Roman (BA) and Daria (Designer), following the new requirements (they will update the docs and Figma) here I delete all the experiments with any kind of statuses except of OPEN. Denis (BE) provides the updated API response here, with the additional 'experimentsAmount' property and I use its value to properly display the total amount of experiments.

* fix: update the types for project.i.ts file

Denis (BE) also updated the 'Projects' for consistancy, so this changes here refers to that