#Ng-387 (#448)
* fix: update the description label name for the 'add project' modal window

* fix: The label texts for the description part for both, 'Add Project' and 'Add Notebook' are properly updated.

Here are changes regarding the placeholer texts of those textareas. There is nothing mentioned about them in the 'Actual result' part, there is only the design screenshot, but on the designs we only see the active state of the textarea. I decide to change it anyway with the copy of text that those appropriate labels include, I think it matches better to the entire form organization. But I ask Daria (designer) about that fact, how they should be displayed, and if she will provide the different direction, then I am going to follow it.